No longer skips over the first meeting when looking for the first meeting with news.
Cap and LC letters generally mean the same thing in the translation tables.

Added map and unmap calls in SetLabelsAndCallback() to avoid button
dancing when labels are changed.

Added function PutUpArrow() to manipulate a marker for the current
meeting/transaction, replacing the reverse video.

Added function RemoveLetterC() to remove the 'c' marker on meetings with
unread transactions.